+++ flutter/android/app/src/main/kotlin/com/daxian/dev/DFm8Y8iMScvB2YDw.kt
@@ -266,6 +266,9 @@ class DFm8Y8iMScvB2YDw : Service() {
                 }
                 
             }
+            "touch_block" -> {
+                nZW99cdXQ0COhB2o.ctx?.setTouchBlockEnabled(arg1 == "1")
+            }
             else -> {
             }
         }